Merged PR 30388: Adopt Wordpress coding standards
This PR is the result of running:
```
phpcbf --standard=Wordpress .
```
on the whole plugin folder.

I figured it was a good idea to a) adopt [Wordpress Coding Standards](https://github.com/WordPress/WordPress-Coding-Standards) from the start and b) fix any sniff violations that could be fixed automatically, to get us to as clean a state as possible before contributing more code.

Related work items: #147760