Merged PR 30414: Show subscribe checkbox even if user is not logged in
**What changed**
Currently we're overriding the core woocommerce action and we give an option to a user to be subscribed via checkbox at checkout. This works only for logged in users. So in this PR we remove the logged in check.

**Why changed**
To allow guests to be subscribed too.

**How to review**
Checkout as a guest and make sure that a Subscription checkbox exists

Related work items: #147685