Merged PR 30762: Pass the callback_url down to the Upgrader
## What's being changed

We're fixing the `wp_remote_post()` call in the `Engagement_Cloud_Upgrader` class.

## Why it's being changed

In this context `$this->callback_url` was not actually set, because we forgot to include it in the new class.

Because we now have a REST callback URL (for EC to call back to WordPress), there was some ambiguity between the new REST callback URL and the existing `$callback_url` (the trackedlink.net URL). We were actually posting a callback_url to a callback_url.

To avoid confusion the original callback_url (which is not really a callback URL at all) has been renamed tracking_url.

## How to review / test this change

- Re-test all the points at which we `wp_remote_post()` to EC
- i.e. activate, upgrade, uninstall

Related work items: #149331