Merged PR 30729: Add settings submenu page with fields
## What's being changed

This PR adds a Settings submenu page for the Engagement Cloud admin section. On the settings page we have two new settings, both of which store values to the `wp_options` table. These settings are used to control the visibility of the marketing checkbox at checkout, and the 'Subscribe to our newsletter' label text presented.

## Why it's being changed

These settings will be used to allow merchants to configure some elements of the plugin's behaviour.

## How to review / test this change

*Test the settings*

- Go the new Settings page
- Confirm the default values are as per the spec
- Save the settings
- Confirm the values stored in wp_options are as intended

*Test the behaviour*

- Delete the options for both settings in the db
- Confirm default behaviours for checkbox visibility and subscribe text
- Uncheck "show checkbox", confirm the checkbox disappears
- Change the text, confirm the text updates

Related work items: #148624