Merged PR 30851: Always run dbDelta to cater for partial upgrades
## What's being changed

We're removing the conditionals that check if a table exists before we create it.

## Why it's being changed

`dbDelta()` is a WordPress method designed to create or update SQL structures. If, for example, there was a partial upgrade, and a merchant had a certain table in their DB but not all the required columns, `dbDelta` will add the missing columns.

## How to review / test this change

- Reset your stored plugin option version to 1.1.0
- Manually drop some columns from your subscriber table e.g.
```
alter table wp_ec_subscribers
drop column updated_at,
drop column created_at
```
- Now refresh the admin
- Your subscribers table should have the missing columns restored

Related work items: #149676