Merged PR 30904: Control checkbox visibility at user registration
## What's being changed

In this PR we added a check to control visibility of the marketing checkbox at user registration form and a new check box field for that purpose.

## Why it's being changed

As it is if you disable "Show marketing checkbox at checkout" you still get the checkbox at user registration page so we wanted to add an extra option for the registration.

## How to review / test this change

- Go to Engagement Cloud > Settings check/uncheck "Show marketing checkbox at register" checkbox
- Confirm that these changes apply to user registration form

Related work items: #149709