Merged PR 30868: Extend use of Engagement_Cloud_Subscriber::create_or_update()
## What's being changed

We've extended the use of this method so it's run in all three places where we handle subscriptions:

- register
- checkout
- widget

I've updated the create_or_update method also - it now sets the created_at and updated_at values for all cases, so we no longer need to set them in the calling code.

## Why it's being changed

In all cases we need to check for a matching subscriber first, and either update or insert depending on the result.

## How to review / test this change

- Sign up via the widget, then register
- Register, then sign up
- Sign up via the widget, then register, then check out, then check out again (with unsubscribe)

Related work items: #149710