Merged PR 30835: Configure unit test environment and add autoloader
**What's being changed**

In this PR we added an autoloader to automatically require all of our classes where a namespace is supplied. We also configure the unit tests environment.

**Why it's being changed**

Using an autoloader is fairly widespread in WordPress plugins, particularly after the codebase reaches a certain size and complexity. Our approach means that when we create a new class, we don't have to add this class to a list of `require_once` expressions in a `load_dependencies()` method.

Being able to reference our classes using `use` statements (as opposed to `require_once()` everywhere) also facilitates unit testing. To this end we've set up a unit testing framework with Brain Monkey and Mockery.

**How to review / test this change**

- composer update on plugin root (this is important so autoload files to be generated)
- put a break pointer in inc/autoloader.php and make sure that all of our classes are being required once you navigate frontend and once you navigate backend
- Test activate- upgrade - uninstall
- Try to break it

**Notes**
2 unit tests added
To run them simply run `vendor/bin/phpunit -c tests/unit/phpunit.xml.dist` on plugin root

Related work items: #148495, #149411