Merged PR 31146: Notify at activation if we didn't already
## What's being changed

At plugin activation, we are notifying EC even if there was no upgrade. This PR also fixes a bug with the plugin activation / deactivation hooks.

## Why it's being changed

It feels safer to do this just in case the information in EC becomes out of sync for some reason. A merchant may deactivate the plugin for a period, then re-activate it - but not need or want to upgrade.

## How to review / test this change

- Deactivate the plugin
- Reactivate plugin
- Confirm that EC was notified

Related work items: #150671, #150677