Merged PR 31403: Fix added_to_cart listener
## What's being changed

We fixed a bug with our front end `added_to_cart` listener.

## Why it's being changed

We were checking for `data` and `data.cart_id` outside of the `cartInsight()` function scope, so if there was no cart insight data when the page loaded, we would exit before the on ready function had a chance to set up its listener.

## How to review / test this change

- Logout as customer
- Clear cookies
- Refresh
- Add to cart from a listing page
- Confirm the network tab shows cart insight data is sent

Related work items: #151555