Merged PR 31262: Identify guests for cart insight tracking
## What's being changed

This PR adds email capture to the plugin. At newsletter sign-up and at guest checkout, we have an `onBlur` event that:
- takes the email address
- posts it to a controller
- returns with an up-to-date cart insight payload
- calls window.dmPt('identify', email)
- sends cart insight

## Why it's being changed

Our client-side adoption of cart insight tracking requires some way of identifying guests. Accordingly we're doing what we've done previously in Magento, i.e. grabbing an email address whenever we see it, and associating that email address with a) the dm_i session (via identify) and b) the Woo session.

We have to store the email address in the Woo session because Cart Insight Handler needs to know if that email address is subscribed, if the option is set to only allow subscribers for AC.

## How to review / test this change

- As a guest, add items to cart
- Sign up for the newsletter
- Check network tab has sent cart insight data
- Re-open an incognito window.
- Repeat the test, this time entering your email address in the billing_email checkout field.
- Turn off engagement_cloud_for_woocommerce_abandoned_cart_allow_non_subscribers
- Repeat the test, no cart insight data should be sent
- Repeat the test in an incognito window with no items added to cart (no data should be sent)

Related work items: #150129