Merged PR 31351: New REST API endpoint to store EC configurations
**What's changed**

In this PR we added a new POST API endpoint to store settings for our plugin. These remote requests will pass the plugin_id for validation, but in addition we are using Woo's own authorization mechanism to validate the REST request via consumer_key and consumer_secret.

**Why it's changed**
Merchants will be saving store settings via Engagement Cloud, and these settings are going to be passed back to WordPress for storage in the options table.

**How to review**
- Create a new consumer_key and consumer_secret in Woocommerce settings > Advanced > REST API.
- Use Postman
- Set the consumer key and secret as the username and password in 'Basic Auth' for the request.
- Create a post request to this endpoint `wp-json/wc-dotdigital-engagement-cloud/v1/configure-store-settings` (1)
- Make sure that we can't send different type of param than it's already set.
- Make sure that the configuration is being saved in wp_options table
- Make sure that the validation works as per keys generated

1. Example POST request uri
```
https://wordpress.ec/wp-json/wc-dotdigital-engagement-cloud/v1/configure-store-settings?non_subscribers=false&cart_delay=23&program_id=12345&wbt_profile_id=DM-123456
```

**Notes**
The transfer of data from EC to Woo requires additional work EC-side to set up the authorization request. The plan is for merchants setting up their plugin connection to approve the authorization request, but this flow will not be active until a future version.

Related work items: #150124