Merged PR 31552: Update module naming conventions
## What's being changed

In this PR we changed all references to 'engagement-cloud', or 'ec' (including code, comments, file names, namespaces) to 'dotdigital', or 'dd'.

## Why it's being changed

The changes are due to a branding move away from Engagement Cloud. We've been really thorough here because a lot of the refs to EC would have been introduced in 1.2.0.

## How to review / test this change

- Make sure that everything works as expected and nothing break in the code
- Make sure no references to engagement-cloud exists in the codebase

Related work items: #151575