Merged PR 31647: Activator test requires table installation
## What's being changed

One of our integration tests. No errors when run in isolation, but when run as part of the suite we were seeing an error about `wp_dotmailer_email_marketing` does not exist.

## Why it's being changed

The Activator test requires the table to exist, because it inevitably runs the upgrade code.

## How to review / test this change

- Run tests

## Notes

We should possibly move the `install_tables` step into a shared setup step. Although in some cases we obviously want to test that the tables are created by the code under test. TBC.

Related work items: #152541