Merged PR 32012: Implement abandoned browse functionality
## What's being changed

In this PR we added a new class to fetch the last browsed product data. Also we modified our existing tracking script to send this to EC, only on product pages.

## Why it's being changed

This main goal of this task is to implement abandoned browse for the Woo plugin. Alongside regular page visits, on product pages we send the current product data to web insight data in EC.

## How to review / test this change

- Make sure that you have wbt profile initialized in wp_options table
- Visit a product page
- Put a break point in web-behaviour-tracking.js and inspect the data
- Check the network tab and review the payload

## Notes

We have not yet synced a brand attribute. Woo doesn't provide one out of the box, and merchants' approach to handling brand is not yet clear.

Related work items: #152160