Merged PR 32103: Do not enqueue scripts if is_admin
## What's being changed

We've added extra logic to prevent the enqueuing of public scripts (and the running of related code) in the WordPress admin.

## Why it's being changed

WordPress's new block-based widget manager (introduced in 5.8) deliberately enqueues public styles and scripts in order to offer a preview of a widget before publishing. This led to a fatal error in our code when we check for an active Woo customer.

## How to review / test this change

- Use WP debug mode
- Add the dotdigital for WooCommerce signup form as a legacy widget
- Ensure no errors are visible on the page

Related work items: #154241