Merged PR 32161: Send full size images for cart insight and web insight
## What's being changed

This PR adapts cart insight and last browsed products to use a shared image helper, with a method to return the original size of the attached image.

## Why it's being changed

Sending thumbnails that are too small in insight data has historically caused problems for merchants in Magento. We're hoping to bypass that here by sending the path to the original image.

## How to review / test this change

- Visit a product page
- Check the JSON sent via `window.dmPt()`
- The image should be the full original upload
- Add product to cart
- Re-check the call to `cartInsight(data)`

Related work items: #154500