Merged PR 32326: Fix WordPress plugin submission feedback
## What's being changed

Our main admin and widget classes.

## Why it's being changed

WordPress had some feedback on our code at submission to the plugins directory. We needed to escape all variables and options when output, and our approach to this wasn't 100% thorough.

Use of the `extract()` method has been removed from the widget class, this is strongly discouraged in the coding standards report (which we were conveniently ignoring).

## How to review / test this change

- Review all the admin settings, check all the controls function as expected
- Review the widget controls in the admin (do all the inputs still work?)
- Review the widget output and function on the front end (does the form still display correctly and submit?)

Related work items: #155124