Merged PR 32183: Remove references to REST API
## What's being changed

In this PR we removed any references to rest-api and callback urls

## Why it's being changed

We've decided, for the time being anyway, not to implement a system whereby configurations are pushed from dotdigital via an authenticated connection. Instead we'll handle configuration on the plugin side.

The REST API class was originally created to provide a controller for unsubscribe requests coming from dotdigital. This also was never actually used - so now there's no need for any sort of REST endpoint in the plugin.

## How to review / test this change

- Review carefully the code and confirm that nothing break

Related work items: #154579