Merged PR 32650: Restore check for program id in cart insight handler
## What's being changed

This PR restores the check for abandoned cart program id in the cart insight handler class.

## Why it's being changed

If there is no program id, then we can't send cart insight - and should exit at this point.

## How to review / test this change

- Clear abandoned cart program id from the admin options
- Leave AC enabled checked
- Ensure no cart insight is sent

## Notes

This does raise the question of whether we need an 'Abandoned cart enabled' checkbox in the admin, if not having a program id is enough to turn it off. However I think we've discussed this and have decided that it's clearer from the merchant's POV.

Related work items: #156294