Merged PR 33491: Check for Woocommerce session when initializing cart
## What's being changed

We're adding some protective checks in case the `woocommerce_set_cart_cookies` is fired earlier than we expect.

## Why it's being changed

A support ticket has been raised that points to a clash with the plugin WooCommerce GlobalE PRO Integration. Evidently this plugin is firing the `woocommerce_set_cart_cookies` hook in the admin, at which point the Woo session has yet to be created. We are treating `woocommerce_set_cart_cookies` as our signal that a cart exists in Woo.

## How to review / test this change

- Run a test case for abandoned cart creation

Related work items: #159120