Merged PR 41509: Handle false or null from wc_get_product()
## What's being changed

We've made some changes to our code wherever we call `wc_get_product()`. This method can return `null` or `false` in addition to a Woo product object.

We're throwing exceptions in the cart-insight script, to be caught in the cart-insight-handler. I did it this way because, in line with our other connectors, a single missing or incorrect line item invalidates the cart insight payload, so we don't send anything.

In class-dotdigital-woocommerce-last-browsed-products.php we have a belt and braces approach. The Throwable catch block does protect against fatal errors if e.g. `get_name()` is called on bool or null, but thought safest to combine this with the if statement.

## Why it's being changed

Merchants have encountered fatal errors from trying to call methods on null.

## How to review / test this change

- Above each of the if statements introduced in this PR, write either `$product = null;` or `$product = false;`
- Switch on all tracking and abandoned cart options in the admin
- Add item to cart
- Test cart insight payload
- Test WBT last product

## Notes

The exceptions thrown in the cart-insight script will be invisible to merchants until we introduce logging of some kind.

Related work items: #192880