Merged PR 46137: Check for WC() customer and session before accessing properties
## What's being changed

In our `Dotdigital_WooCommerce_Customer` class we're adding some extra checks to ensure we aren't calling methods on null.

## Why it's being changed

We have seen this error:
```
PHP Fatal error:  Uncaught Error: Call to a member function get_email() on null in <root>/plugins/dotdigital-for-woocommerce/includes/customer/class-dotdigital-woocommerce-customer.php:28
```

It's proven tricky to reproduce this reliably, but it must be possible in some cases for there to be a current user id while `WC()->customer` is null.

## How to review / test this change

- In Woo, switch on tracking and AC for all users including non-subscribers
- Log in as a customer and add an item to the cart
- Confirm the identify call in the Network tab shows your logged in user's email
- Now log out
- Add a 'Dotdigital for WooCommerce Signup Form' widget to a widgetable area on your site
- Put an email address into the input on the storefront, and trigger a blur to capture the guest email
- See the identify call shows your guest email

Related work items: #216274