Merged PR 36154: Set up ESLint for the plugin JS
## What's being changed

I've added some dev dependencies for running ESLint on our JS.

## Why it's being changed

As we write more JS for the plugin, it's probably sensible to align to an agreed coding style. This PR suggests a baseline, but we can review how obtrusive / onerous this is and adjust as required.

Related work items: #169508