Merged PR 59378: Protect against get_the_terms returning false
## What's being changed

This handles the case that `get_the_terms()` returns false.

## Why it's being changed

If a product in the cart has no terms for 'product_cat', we see a PHP warning.

Note I could not replicate this because in my installation, removing categories from a product sets the Category to 'Uncategorised'.

## How to review / test this change
- Configure cart insight in the plugin settings
- In includes/category/class-dotdigital-woocommerce-category.php, set $wc_terms to false
- Add a product to the cart
- On the storefront, see no error
- Inspect the cart insight payload - it should show empty string for 'category' under each line item
- Check the payload is received in Dotdigital (you will have to log in first to have the data associated with a contact)

Related work items: #275841